emitter api
