+++ src/aioway/io/__init__.py
@@ -4,4 +4,5 @@ from ._media import *
 from .dsets import *
 from .indices import *
 from .sess import *
+from .sinks import *
 from .volatile import *
