+++ src/aioway/io/dsets.py
@@ -21,7 +21,6 @@ __all__ = [
     "TdictLoaderHop",
     "Dset",
     "Stream",
-    "Sink",
     "Frame",
     "dset_dcls",
     "TensorStream",
@@ -186,36 +185,6 @@ class TdictStream(_TdictAttrsMixin, Stream[td.TensorDict], abc.ABC):
     """
     A `TdictStream` is a `Stream` of `torch.Tensor`s.
     """
-
-
-@dset_dcls
-class Sink[T = typing.Any](Dset, abc.ABC):
-    """
-    Consumes a `Hop` and writes to some external location.
-    """
-
-    TYPE: typing.ClassVar[type[T]]
-    """
-    The type to check
-    """
-
-    def __call__(self, hop: Hop[T]) -> None:
-        for batch in hop:
-            if not isinstance(batch, self.TYPE):
-                raise TypeError(f"The batch has {type(batch)=}, expected {self.TYPE}.")
-
-            self.write(batch)
-
-    @abc.abstractmethod
-    def write(self, batch: T) -> None:
-        raise NotImplementedError
-
-    @typing.override
-    def _register(self) -> None:
-        from .sess import SinkSession
-
-        if sess := SinkSession.current():
-            sess.push(self)
 
 
 @dset_dcls
