+++ src/aioway/io/sess.py
@@ -8,7 +8,8 @@ import typing
 from aioway._utils import Stack
 from aioway.sess import Session
 
-from .dsets import Dset, Sink
+from .dsets import Dset
+from .sinks import Sink
 
 __all__ = ["DsetSession", "SinkSession"]
